1.2.1 - 可以查看新版课表

+++ app/src/main/java/zz/itcast/studentschedule/db/ExcelDao.java
@@ -408,7 +408,9 @@ public class ExcelDao {
     }
 
     public void clearData() {
-        db.delete(table_ss,null,null);
+        if(db!=null){
+            db.delete(table_ss,null,null);
+        }
         instance = null;
     }
